Implement Macintosh functions CopyBits and CopyMask
CopyMask isn't quite complete: the mask is allowed to be a grayscale
image. For a grayscale mask, if the color of a pixel is between black
and white, then the corresponding source and destination pixels are
blended together. The mask images in the classic Macintosh version of
Glider PRO don't utilize this feature, but the mask images in the
Mac OS X port do.