Partially implement function HandleEvent
When mapping the Macintosh controls to Windows, I have chosen to map
the COMMAND key to the Ctrl key and the OPTION key to the Alt key.
The SHIFT key continues to be the Shift key. This may be tweaked if
it conflicts with other standard controls (for example, Alt+<letter>
in dialog box navigation).