Adjust oval drawing functions to touch rectangle
This change makes the oval functions draw their shapes so that the
outside right and bottom edges of the oval touches the inside right
and bottom edges of the given bounding rectangle. Before this change,
the ColorOval and ColorShadowOval functions touched the top and left
edges, but were one pixel away from the bottom and right edges.
(ColorFrameOval used Ellipse instead of FillRgn, so it worked ok.)
The elliptic regions are now created a tad larger, to compensate,
and ColorFrameOval uses FrameRgn now.