BUGFIX: Check array bounds in GetNumberOfLights
The value of 'where' is an index into the room array, usually
retrieved from the 'localNumbers' array. The 'localNumbers' array
stores the room numbers of rooms that surround the central room,
or (-1) if the room doesn't exist in the house file. This (-1)
value was not checked for in GetNumberOfLights, but now it will be.