BUGFIX: Initialize hotspot number for all grease
Two factors combine to form a potential access violation:

(1) When the player moves to a room, the game takes note of all
standing jars of grease by calling AddGrease() when it draws them.
AddGrease() didn't initialize the hotNum field. The hotNum field was
initialized later, when a glider knocks over a jar of grease.

(2) When a glider collects a bonus item, the spilled lines of grease
are redrawn, to handle the case where the bonus item was sitting over
the grease spill. To do this, the game grabs the grease line's
rectangle with "hotSpots[grease[i].hotNum].bounds", and then tests
whether the grease line should be drawn.

So: when there is a standing jar of grease visible to the player, and
the player collects a bonus item, the game indexes into the hotSpots
array with a garbage value (either uninitialized, or left over from
a previous room's state).

To fix this, the hotNum field is initialized to -1 when jars of grease
are added to be tracked, and if the hotNum isn't in range when the
grease lines are redrawn, the correspond grease jar is skipped.