Implement load house dialog
Most of the custom code that was present is now replaced by the
Windows list-view control. The list-view control does everything that
the custom Macintosh event handling did:

(1) If the user double-clicks on a house, open that house.
(2) If the user types a letter, select the first house beginning with
    that letter.
(3) If the user presses arrow keys, then move the selection to another
    house.
(4) If the user presses the PAGE UP or PAGE DOWN key, then move the
    selection up or down by one page.

For this and other dialogs, the Windows standard dialog behavior will
replace the corresponding code that did these things:

(1) If the user presses the ENTER key, then act as if the OK button
    was clicked.
(2) If the user presses the ESC key, then act as if the Cancel button
    was clicked.