BUGFIX: Prevent out of bounds access to 'pages'
The index into 'pages', 'which', is calculated as follows:

 if ((angelDest.left % 32) == 0)
 {
  /*...*/
  which = angelDest.left / 32;
  which = which % 5;
  ZeroRectCorner(&pages[which].dest);
  /*...*/
 }

The line 'which = which % 5;' sets 'which' to a negative value if
'angelDest.left' is negative. This negative index then performs the
out of bounds access.

The entire block doesn't need to be performed while the flying angel
is partially off of the left side of the screen. This is because the
stars that she drops won't be visible until she is fully visible. The
fix is to skip the block if the angel's left side is less than zero
(i.e., off the screen). That way, 'angelDest.left' and 'which' will
always be positive.