Use GetAsyncKeyState instead of GetKeyState during game
GetKeyState is only really meant to be called from somewhere within a
window procedure, as far as I can tell. The return values of
GetKeyState were sometimes strange, resulting in the game window
reacting to keystrokes even if the window was inactive. To mitigate
this, GetAsyncKeyState is used instead during the gameplay, wrapped in
macros that also check whether the main window is the currently active
window. If the main window is not active, then all keys are considered
to be up.

  #define IsKeyDown(vkey) \
    (GetActiveWindow() == mainWindow && GetAsyncKeyState(vkey) < 0)

  #define IsKeyUp(vkey) \
    (GetActiveWindow() != mainWindow || GetAsyncKeyState(vkey) >= 0)