Remove 'doBackground' loop in PlayGame()
This loop just called HandlePlayEvent, which was never implemented
directly and has now been superseded by the WM_ACTIVATEAPP handling
in MainWindowProc(). If 'doBackground' was true (i.e., the Background
Tasks checkbox was checked in Brains Preferences), then when the game
was deactivated, it would fall into an infinite loop. Removing this
useless loop removes this problem entirely.

This was the only place that the Background Tasks preference was ever
utilized, so the option will be removed entirely at a later time.