Fix up music mode mistakes
Before this patch, my solution for queuing up the buffers of sound data
for playback was flawed. The code was queuing up new sound to play as
soon as possible, but that meant that the queued sound actually began
playing some time later.

This new solution only figures out which music piece to queue next
if that music piece is actually going to play in the next 50ms (the
length of a music tick), instead of setting it to play a couple of
seconds from the current time.