BUGFIX: Properly handle doors in AddObjectPairing
When AddObjectPairing is called and the active object is an exterior
door object, then the game checks whether it should add a matching
interior door object in the next room. However, it checks for the
wrong thing:

 // checking for interior doors in next room (WRONG)
 (!DoesRoomNumHaveObject(roomNum, kDoorInLfLeft))
 (!DoesRoomNumHaveObject(roomNum, kDoorInRtLeft))

The constants kDoorInLfLeft and kDoorInRtLeft are actually coordinates
for where the door objects should be placed in the room. The correct
parameters are the interior door object IDs, kDoorInLf and kDoorInRt:

 // checking for interior doors in next room (RIGHT)
 (!DoesRoomNumHaveObject(roomNum, kDoorInLf))
 (!DoesRoomNumHaveObject(roomNum, kDoorInRt))