BUGFIX: merge floor suite properly in v1 houses
Before, if the user loaded up an old version 1.0 house and changed
or added any links between objects, those links would become messed
up. This is because the room and suite numbers were packed together
by MergeFloorSuite for a version 2.0 house. Later, when the house is
saved and converted to a version 2.0 house, the packed version 2.0
links are interpreted as packed version 1.0 links and repacked as
version 2.0 links. This completely mangles where the links point
to within the house.

To fix this, MergeFloorSuite now merges the floor and suite together
properly for version 1.0 houses. It checks the house version like
ExtractFloorSuite does in order to do the right thing. When converting
a version 1.0 house to version 2.0, a new function (MergeFloorSuiteVer2)
is called to explicitly pack the floor and suite together for the
version 2.0 house.