BUGFIX: don't use uninitialized rect in HandleSwitches
The bounds rectangle for an AddSparkle call was never initialized.
This is likely because the AddSparkle call *was* used in the past,
before the RestoreFromSavedMap function gained the doSparkle parameter.
This parameter is set to true in a call to RestoreFromSavedMap directly
before the bad AddSparkle call, so the AddSparkle call can simply be
removed.