resfork-extract: Add FFmpeg's MACE audio decoder
This removes the external dependency on the FFmpeg command-line tool
to convert the .AIFF files to .WAV files, because the only reason that
was needed was because the sound data is sometimes compressed with
MACE 3-to-1 or MACE 6-to-1 compression. (MACE stands for Macintosh
Audio Compression and Expansion.) Now .WAV files can be outputted
directly with less mess and less fuss.