Add new resource loading interface
The idea is put all resource loading functions in one place, and to
allow the transition to using zip files for storing built-in game
assets and house data. In this initial commit, built-in images are
loaded from Mermaid.dat, but built-in sounds and house resources are
still loaded in the old way. This will be rectified in the following
commits.