Don't check for negative dimensions in QInsetRect
This fixes some rendering bugs with the very thin furniture used in
the SpacePods house. This house relies on the object drawing code's
odd behavior when drawing slim furniture, including the fact that
rectangles with negative width or height happen during the drawing
process. It expects the negative width and height to happen.