BUGFIX: do proper outlet collision in two-player mode
To reproduce in a version without this patch:

(1) Create a house with a single room containing an always-on
    electrical outlet. To make the rest of the procedure quicker,
    set the interval between zaps to be small.
(2) Start a new two-player game with this house.
(3) Repeatly kill the gliders until only a single glider is left on
    the screen. You can do this by killing one glider until it never
    comes back, or by taking turns until someone doesn't come back.
(4) With the single glider, pass over the zapping outlet. Observe
    that it doesn't kill the glider (unless your viewport resolution
    is very small).

The reason this happens is because the dynamic object for an electrical
outlet stores its destination rectangle relative to the top-left corner
of `workSrcMap`, but calls CheckDynamicCollision() with parameters that
state that the destination rectangle is relative to the top-left corner
of the center room (the wrong thing to do). The wrong check is only
performed when the game is two-player and there is only one player left
on the screen.

The fix is to just fix up the parameter values, setting the `doOffset`
parameter to `true` in all cases.