BUGFIX: fix stars freezing in game-over animation
When the player finishes the game by collecting all of the stars in the
house, an animation plays where a glider drops stars down the screen to
reveal the house trailer message. If the screen is very large, near the
maximum resolution of 1536 by 986 pixels, then most of the stars will
freeze in place before disappearing off of the screen.

This bug is caused by the way the star animation is controlled: when a
new star is dropped as the glider moves right, the leftmost star stops
animating. On smaller resolutions, this isn't noticed because the star
which stops animating stops off of the screen. The solution here is to
allow more stars to be animating at a time (from 5 stars to 8 stars),
and to increase the delay before stopping the animation altogether.