Always use pure black dithered shadows
The color values used when calling the ColorShadow functions were only
indexes that resolved to #111111 (very dark gray) or #000000 (black).
#111111 was used for object shadows, and #000000 was used to shadow
dark rooms in the editor mode. The difference between the two is only
visible when compared next to each other. Restricting the dithered
shadows to pure black allows the rendering code for the shadows to
take about half the time to run (because there is only one FillRgn()
call instead of two).