Restore menu bar after game over animations
Showing the scoreboard during the game over animations is more fitting
than showing a disabled menu bar. Also, if the menu bar height is not
exactly 20 pixels, then the game's pixels are overlapped or shifted
down at the top by the menu bar (depending on whether UpdateWindow()
is called).

This change also fixes the visual bug of erasing moving objects when
the glider uses all of its lives and the died-game-over animation
happens.