Rewrite cursor handling code
The HideCursor calls were always disabled. Their purpose was to
hide the cursor during gameplay; this isn't necessary. If the player
doesn't want to see their cursor in the game area, they can move it
aside themselves.

SpinCursor was called to show a wait cursor, but everything happens
so fast that the player never sees it for more than a split second.
If some operation does end up being slow, a call like

    SetCursor(LoadCursor(NULL, IDC_ARROW));

will suffice.

InitCursor was called to switch a wait cursor back to arrow, or to
restore the arrow cursor after a drag operation. These have been
removed or replaced as appropriate.

When system cursors are needed (say, for drag operations), they will
be loaded up as needed, instead of pre-loading them at startup.