Reimplement dissolve transition code
This version properly times itself to a specific frame rate instead of
going as fast as possible. In addition, it runs much faster than the
previous implementation did. Using DIB sections to gain direct access
to the temporary bitmap memory side-stepped the overhead of calling
BitBlt thousands of times for small pixel chunks.