Fix interaction oddities with WM_CLOSE message
Before this patch, you could click the close button on the main window
before the game begins (while an opening banner is being shown) or
after the game ends (caused by death or triumph). The game asked if
you wanted to save the game at those times, but it doesn't make sense.
Now the game checks for these cases and doesn't bother asking if you
you want the game to be saved. In one case, the game hasn't begun yet,
and in the other case, the game has finished and there is nothing left
to save.