Add support for custom splash screens
This patch is careful to make sure that the splash screen displayed
after the demo game finishes is always the splash screen for the
previous house. It achieves this with an extra parameter to ReadHouse:
the 'loadSplashScreen' parameter. Typically, this parameter is set to
'true' to tell ReadHouse to update the splash screen image to match
whatever is appropriate for the house being read. However, when the
demo house is temporarily loaded for the automatic demo game, the
game shouldn't update the splash screen. This is because, if it did
update the splash screen then, the game would display the wrong splash
screen during the dissolve transition that happens after the demo
finishes. To handle this corner case properly, the 'loadSplashScreen'
parameter is set to 'false' when calling ReadHouse for the demo game.

A future update will remove the need for a previously added hack that
was added to handle the above-mentioned corner case: the
'splashHouseIndex' parameter in the PlayGame and NewGame functions.
It will do this by precomposing the entire splash screen at house load
time, instead of calling DrawOnSplash during each redraw later on
(including the redraw in the RedrawSplashScreen call in PlayGame).