Allow glider suicide with Backspace & Delete keys
Macintosh keyboards have a key labeled "Delete" which functions like
"Backspace" on a Windows keyboard. To avoid confusion, allow both
Backspace and Delete keys for player input to destroy the straggling
glider in two-player mode. (This change has already been made for
deleting objects and rooms in the house editor.)