resfork-extract: fixed 16-bit image conversion bug
This fixes a bug where the red channel of a 16-bit image could be wrong
in the lowest bit after converting to Windows BMP format. The mask
is changed from the value 0x7A (01111010 in binary) to the correct
value 0x7C (01111100 in binary).

For an example, check PICT 3001 in the resource fork of the house
"The Asylum Pro" (in the Houses folder).