Restore colored dithered shadow functions
Back in commit 2892b1da, "Always use pure black dithered shadows", I decided to remove support for the dark gray shadows. Not sure why I did that, besides the FillRgn calls changing from 2 calls to 1 call. I'm restoring the colored shadows here to better match what the game is meant to look like. In the future, GDI's AlphaBlend() function may be useful for a fancier shadow effect.

The original effect was done with a pen pattern and the "patOr" pen mode. The closest GDI equivalent, taking into account the color differences between GDI and QuickDraw, is a pattern brush with the "R2_MASKPEN" mixing mode. Instead of the monochrome pattern shenanigans, I'll use a DIB pattern brush with real colors.

Also, I'm adding a CreateColorShadowBrush() function to help out with this. This function is also used for the editor's map window and its dithered colored boxes.